traits
Lots of redundant code around for handling things like getting/setting
properties and iterating objects that use the internal properties
arrays. Set these up as traits.
